Blazor

+++ DropSpace.Contracts/Dtos/FileModelDto.cs
@@ -1,4 +1,4 @@
 ﻿namespace DropSpace.Contracts.Dtos
 {
-    public record class FileModelDto(Guid Id, double Size, double SizeMb, string FileName, PendingUploadModelDto Upload);
+    public record class FileModelDto(Guid Id, long Size, double SizeMb, string FileName, PendingUploadModelDto Upload);
 }
